Move ``not-callable`` from ``TestTypeChecker`` unittests
